requirements.txt
streamlit>=1.32.0
pymysql>=1.1.0
pandas==2.1.4
plotly>=5.20.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
xlwt>=1.3.0
openpyxl>=3.1.0
scikit-learn>=1.4.0
numpy==1.26.4
joblib>=1.3.0
python-dateutil>=2.9.0
